add interface definition(s) for test details
