+++ src/app/app.module.ts
@@ -5,8 +5,8 @@ import { NgModule } from '@angular/core';
 import { AppComponent } from './app.component';
 import { TestExecDetailsModule } from './modules/details/test-exec-details.module';
 import { MessagingModule } from '@testeditor/messaging-service';
-import { TestExecutionDetailsService } from './modules/test-execution-details/test-execution-details.service';
-import { DefaultTestExecutionDetailsService } from './modules/test-execution-details/test-execution-details.service';
+import { TestExecutionDetailsService } from './modules/details-service/test-execution-details.service';
+import { DefaultTestExecutionDetailsService } from './modules/details-service/test-execution-details.service';
 
 @NgModule({
   declarations: [
