--- src/app/modules/test-execution-details/test-execution-details.service.ts
@@ -1,23 +1,0 @@
-import { Injectable } from '@angular/core';
-
-export interface TestExecutionDetails {
-
-}
-
-export abstract class TestExecutionDetailsService {
-  abstract getTestExecutionDetails(
-    jobID: number,
-    onResponse?: (details: TestExecutionDetails) => void,
-    onError?: (error: any) => void
-  ): void;
-}
-
-@Injectable()
-export class DefaultTestExecutionDetailsService extends TestExecutionDetailsService {
-
-  getTestExecutionDetails(jobID: number, onResponse?: (details: TestExecutionDetails) => void, onError?: (error: any) => void): void {
-    throw new Error('Method not implemented.');
-  }
-  constructor() { super(); }
-
-}
